organized accountManager, I think this organization is good and doesn't require any more major refactoring. It would probably help to improve some names, but it's not urgent
